DTSCCI-5034: Update Idam endpoints in setup (#7697)
* DTSCCI-5034: Update Idam endpoints

+++ bin/utils/ccd-import-definition.sh
@@ -9,8 +9,17 @@ filepath=${1}
 filename=$(basename ${filepath})
 uploadFilename="$(date +"%Y%m%d-%H%M%S")-${filename}"
 
-userToken=$(${dir}/idam-lease-user-token.sh ${CCD_CONFIGURER_IMPORTER_USERNAME:-ccd.docker.default@hmcts.net} ${CCD_CONFIGURER_IMPORTER_PASSWORD:-Password12!})
-serviceToken=$(${dir}/idam-lease-service-token.sh ccd_gw $(docker run --rm hmctspublic.azurecr.io/imported/toolbelt/oathtool --totp -b ${CCD_API_GATEWAY_S2S_SECRET:-AAAAAAAAAAAAAAAC}))
+if [ -z "${USER_TOKEN:-}" ]; then
+  userToken=$(${dir}/idam-lease-user-token.sh ${CCD_CONFIGURER_IMPORTER_USERNAME} ${CCD_CONFIGURER_IMPORTER_PASSWORD})
+else
+  userToken=${USER_TOKEN}
+fi
+
+if [ -z "${SERVICE_TOKEN:-}" ]; then
+  serviceToken=$(${dir}/idam-lease-service-token.sh ccd_gw $(docker run --rm hmctspublic.azurecr.io/imported/toolbelt/oathtool --totp -b ${CCD_API_GATEWAY_S2S_SECRET:-AAAAAAAAAAAAAAAC}))
+else
+  serviceToken=${SERVICE_TOKEN}
+fi
 
 version="n/a"
 newVersion="n/a"
